#76 Update to template
