+++ .github/workflows/add-dependabot-request-to-project-t-comm.yaml
@@ -2,14 +2,17 @@
 # - https://github.com/srggrs/assign-one-project-github-action
 
 name: add-dependabot-request-to-project-t-comm.yaml
+
 on:
   pull_request:
     types:
       - opened
       - reopened
+
 env:
   GITHUB_TOKEN: ${{ secrets.SENZING_GITHUB_ACCESS_TOKEN }}
   CREATOR: ${{ github.event.pull_request.user.login }}
+
 jobs:
   add-dependabot-request-to-project:
     name: Add dependabot pull request to Senzing Community project
