+++ .github/workflows/gosec.yaml
@@ -2,6 +2,7 @@
 # - https://github.com/securego/gosec
 
 name: gosec.yaml
+
 on:
   push:
     branches:
@@ -9,6 +10,7 @@ on:
   pull_request:
     branches:
       - main
+
 jobs:
   gosec:
     runs-on: ubuntu-latest
