+++ .github/workflows/identify-customer.yaml
@@ -1,12 +1,15 @@
 name: identify-customer.yaml
+
 on:
   issues:
     types:
       - opened
+
 env:
   GITHUB_TOKEN: ${{ secrets.ORG_MEMBERSHIP_TOKEN }}
   MEMBER_LIST: ${{ secrets.SENZING_MEMBERS }}
   CREATOR: ${{ github.actor }}
+
 jobs:
   identify-customer:
     name: Add customer-submission label
